Refactor: Move account validator to a separate module
