+++ exceptions/account_invalid_error.rb
@@ -1,0 +1,5 @@
+class AccountInvalidError < StandardError
+  def initialize(msg="InvalidAccount")
+    super
+  end
+end
